// frontend/src/app/core/api/generated/model/taskResponse.ts
/**
 * Bla.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaskStatus } from './taskStatus';


export interface TaskResponse { 
    id: string;
    title: string;
    description: string | null;
    status: TaskStatus;
    dueDate: string;
    createdAt: string;
    updatedAt: string;
}
export namespace TaskResponse {
}