// frontend/src/app/core/api/generated/model/taskStatus.ts
/**
 * Bla.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const TaskStatus = {
    Todo: 'Todo',
    InProgress: 'InProgress',
    Done: 'Done'
} as const;
export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];